All new faceted browsing implementation.
Any operation like adding a filter or performing a collection transformation (e.g. groupBy) is wrapped in a Command now. Commands are reversible. The state is stored in memory, so fast undo/redo actions are possible. Commands are ready to be persisted on the server, so sheets can be stored and reproduced by executing the command sequence. No data needs to be stored, just logic. Pretty cool! :)